Fix color transparency issues and improve settings synchronization
- Convert color representation from Long to Int in all models (LiquidGlassSettings, LauncherItem).
- Use it.toArgb() in color pickers to ensure correct 32-bit ARGB values are stored.
- Fix default color values to use proper Int representation.
- Add "Secret Wallpaper" feature to debug settings with improved state synchronization.
- Fix launcher closing issue by refining BackHandler on the home screen.
- Restore theme-aware Day/Night wallpaper support.
- Add Notification Dots with customizable color and liquid glass style.

+++ app/src/main/kotlin/com/quimodotcom/lqlauncher/activities/LiquidGlassLauncherActivity.kt
@@ -2621,7 +2621,7 @@ private fun createDefaultItems(apps: List<AvailableApp>): List<LauncherItem> {
             spanY = 2,
             panelType = PanelType.CLOCK,
             blurRadius = 25f,
-            tintColor = 0xFF6366F1
+            tintColor = 0xFF6366F1.toInt()
         )
     )
