+++ app/src/main/kotlin/com/quimodotcom/lqlauncher/compose/launcher/LiquidGlassSettings.kt
@@ -28,7 +28,7 @@ data class LiquidGlassSettings(
     val vibrancyEnabled: Boolean = true,
     
     // Colors and transparency
-    val panelTintColor: Long = 0xFF6366F1,
+    val panelTintColor: Int = 0xFF6366F1.toInt(),
     val panelBackgroundAlpha: Float = 0.12f,
     val iconBackgroundAlpha: Float = 0.1f,
     
@@ -85,7 +85,7 @@ data class LiquidGlassSettings(
 
     // Notifications
     val showNotificationDots: Boolean = false,
-    val notificationDotColor: Long = 0xFFEF4444, // Default Red-500
+    val notificationDotColor: Int = 0xFFEF4444.toInt(), // Default Red-500
     val liquidGlassNotificationDots: Boolean = false
 )
